funciona pero tengo error en el controlador de Upload

+++ Pae.Web/Pae.web/Pae.web/Data/Entities/Estudents.cs
@@ -10,7 +10,7 @@ namespace Pae.web.Data.Entities
 
         [Required(ErrorMessage ="El campo {} es requerido")]
         [Display(Name ="Documento")]
-        public int Document { get; set; }
+        public long Document { get; set; }
 
         [Required(ErrorMessage = "El campo {} es requerido")]
         [Display(Name = "Nombre Completo")]
